# dashboard/.schema-openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Orchestrator Dashboard",
        "description": "REST/SSE API of the YAOE-FLOW observability dashboard.",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/webhooks": {
            "get": {
                "operationId": "getApiWebhooks",
                "tags": [
                    "Webhooks"
                ],
                "summary": "List webhook events with filters and pagination",
                "responses": {
                    "200": {
                        "description": "Webhooks",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "propertyNames": {
                                        "type": "string"
                                    },
                                    "additionalProperties": {}
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "in": "query",
                        "name": "issueId",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "in": "query",
                        "name": "teamId",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "in": "query",
                        "name": "projectId",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "in": "query",
                        "name": "q",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "in": "query",
                        "name": "page",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "in": "query",
                        "name": "pageSize",
                        "schema": {
                            "type": "number"
                        }
                    }
                ]
            }
        },
...
        "/api/webhooks/{id}": {
            "get": {
                "operationId": "getApiWebhooksById",
                "tags": [
                    "Webhooks"
                ],
                "summary": "Webhook event detail",
                "responses": {
                    "200": {
                        "description": "Webhook encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "propertyNames": {
                                        "type": "string"
                                    },
                                    "additionalProperties": {}
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "N\u00e3o encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "error"
                                    ]
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "schema": {
                            "type": "string"
                        },
                        "in": "path",
                        "name": "id",
                        "required": true
                    }
                ]
            }
        },
...
        "/api/users": {
            "get": {
                "operationId": "getApiUsers",
                "tags": [
                    "Users"
                ],
                "summary": "List users",
                "responses": {
                    "200": {
                        "description": "Usu\u00e1rios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "propertyNames": {
                                        "type": "string"
                                    },
                                    "additionalProperties": {}
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "operationId": "postApiUsers",
                "tags": [
                    "Users"
                ],
                "summary": "Create user",
                "responses": {
                    "200": {
                        "description": "Criado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "const": true
                                        },
                                        "user": {
                                            "type": "object",
                                            "propertyNames": {
                                                "type": "string"
                                            },
                                            "additionalProperties": {}
                                        }
                                    },
                                    "required": [
                                        "ok",
                                        "user"
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de valida\u00e7\u00e3o",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "error"
                                    ]
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "anyOf": [
                                            {
                                                "type": "string"
                                            },
                                            {
                                                "type": "null"
                                            }
                                        ]
                                    },
                                    "username": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "name",
                                    "username",
                                    "password"
                                ]
                            }
                        }
                    }
                }
            }
        },
        "/api/users/{id}": {
            "patch": {
                "operationId": "patchApiUsersById",
                "tags": [
                    "Users"
                ],
                "summary": "Update user",
                "responses": {
                    "200": {
                        "description": "Atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "const": true
                                        },
                                        "user": {
                                            "type": "object",
                                            "propertyNames": {
                                                "type": "string"
                                            },
                                            "additionalProperties": {}
                                        }
                                    },
                                    "required": [
                                        "ok",
                                        "user"
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "error"
                                    ]
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "string",
                            "minLength": 1
                        },
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "anyOf": [
                                            {
                                                "type": "string"
                                            },
                                            {
                                                "type": "null"
                                            }
                                        ]
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "inactive"
                                        ]
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/profile/password": {
            "post": {
                "operationId": "postApiProfilePassword",
                "tags": [
                    "Users"
                ],
                "summary": "Change your own password (requires the current password)",
                "responses": {
                    "200": {
                        "description": "Senha alterada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "const": true
                                        }
                                    },
                                    "required": [
                                        "ok"
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de valida\u00e7\u00e3o",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "error"
                                    ]
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "currentPassword": {
                                        "type": "string"
                                    },
                                    "newPassword": {
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "currentPassword",
                                    "newPassword"
                                ]
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}